+++ src/Rdd.Infra/Storage/IStorageService.cs
@@ -7,7 +7,10 @@ namespace Rdd.Infra.Storage
     public interface IStorageService
     {
         IQueryable<TEntity> Set<TEntity>() where TEntity : class;
+
         Task<IEnumerable<TEntity>> EnumerateEntitiesAsync<TEntity>(IQueryable<TEntity> entities) where TEntity : class;
+        Task<int> CountAsync<TEntity>(IQueryable<TEntity> entities) where TEntity : class;
+
         void Add<TEntity>(TEntity entity) where TEntity : class;
         void AddRange<TEntity>(IEnumerable<TEntity> entities) where TEntity : class;
         void Remove<TEntity>(TEntity entity) where TEntity : class;